Refactor styles
Changed navigation styles - added active link, theme button moved to the
bottom of nav and fixed the position of the nav.
File-box class was renamed to card-wrapper.
Added additional classes to help card styling on index page.
Changed card styles - responsive needs to be fixed.
Added logo to the header.